requirements.txt
opencv-python
opencv-contrib-python
mysql-connector
numpy
pillow